make single script for PowerNVEEH.py and PowerVMEEH.py
add EEH.py test by combine both PowerNVEEH.py and PowerVMEEH.py test.
