+++ io/pci/EEH.py
@@ -25,6 +25,7 @@ from avocado.utils import process
 from avocado.utils import pci
 from avocado.utils import genio
 from avocado.utils import distro
+from avocado.utils import dmesg
 from avocado.utils.software_manager import SoftwareManager
 
 EEH_HIT = 0
@@ -32,20 +33,21 @@ EEH_MISS = 1
 
 
 class EEHRecoveryFailed(Exception):
+
     """
     Exception class, if EEH fails to recover
     """
 
-    def __init__(self, msg, dev, log=None):
-        self.msg = msg
+    def __init__(self, thing, dev, log=None):
+        self.thing = thing
         self.dev = dev
         self.log = log
 
     def __str__(self):
-        return "%s %s recovery failed: %s" % (self.msg, self.dev, self.log)
+        return "%s %s recovery failed: %s" % (self.thing, self.dev, self.log)
 
 
-class PowerNVEEH(Test):
+class EEH(Test):
 
     """
     This class contains functions for listing domains
@@ -58,37 +60,42 @@ class PowerNVEEH(Test):
         """
         if 'ppc' not in distro.detect().arch:
             self.cancel("Processor is not ppc64")
-        output = genio.read_file("/sys/kernel/debug/powerpc/eeh_enable")\
-            .strip()
-        if output != '0x1':
+        eeh_enable_file = "/sys/kernel/debug/powerpc/eeh_enable"
+        if '0x1' not in genio.read_file(eeh_enable_file).strip():
             self.cancel("EEH is not enabled, please enable via FSP")
-        self.max_freeze = int(self.params.get('max_freeze', default='1'))
-        cmd = "echo %d > /sys/kernel/debug/powerpc/eeh_max_freezes"\
-            % self.max_freeze
-        process.system(cmd, ignore_status=True, shell=True)
-        self.function = str(self.params.get('function', default='4'))
-        self.err = str(self.params.get('err'))
-        self.pci_device = str(self.params.get('pci_device', default=""))
+        self.max_freeze = self.params.get('max_freeze', default=1)
+        self.pci_device = self.params.get('pci_device', default="")
+        self.add_cmd = self.params.get('additional_command', default='')
         if not self.pci_device:
-            self.cancel("PCI bus id not given")
-        self.phb = self.pci_device.split(":", 1)[0]
-        self.addr = genio.read_file("/sys/bus/pci/devices/%s/"
-                                    "eeh_pe_config_addr" % self.pci_device)
-        self.addr = str(self.addr).rstrip()
-        self.err = 0
+            self.cancel("No PCI Device specified")
+        self.function = str(self.params.get('function')).split(" ")
         smm = SoftwareManager()
         if not smm.check_installed("pciutils") and not smm.install("pciutils"):
             self.cancel("pciutils package is need to test")
-        for line in process.system_output('lspci -vs %s' % self.pci_device,
-                                          ignore_status=True,
-                                          shell=True).decode("utf-8\
-                                          ").splitlines():
-            if 'Memory' in line and '64-bit, prefetchable' in line:
-                self.err = 1
-                break
         self.mem_addr = pci.get_memory_address(self.pci_device)
         self.mask = pci.get_mask(self.pci_device)
-        self.log.info("Test-----> %s", self.addr)
+        if self.is_baremetal():
+            cmd = "echo %d > /sys/kernel/debug/powerpc/eeh_max_freezes"\
+                % self.max_freeze
+            process.system(cmd, ignore_status=True, shell=True)
+            self.phb = self.pci_device.split(":", 1)[0]
+            self.addr = genio.read_file("/sys/bus/pci/devices/%s/"
+                                        "eeh_pe_config_addr" % self.pci_device)
+            self.addr = str(self.addr).rstrip()
+            self.err = 0
+            for line in process.system_output('lspci -vs %s' % self.pci_device,
+                                              ignore_status=True,
+                                              shell=True).decode("utf-8\
+                                              ").splitlines():
+                if 'Memory' in line and '64-bit, prefetchable' in line:
+                    self.err = 1
+                    break
+        else:
+            self.pci_class_name = pci.get_pci_class_name(self.pci_device)
+            if self.pci_class_name == 'fc_host':
+                self.pci_class_name = 'scsi_host'
+            self.pci_interface = pci.get_interfaces_in_pci_address(
+                self.pci_device, self.pci_class_name)[-1]
         self.log.info("===============Testing EEH Frozen PE==================")
 
     def test_eeh_basic_pe(self):
@@ -103,7 +110,15 @@ class PowerNVEEH(Test):
                 self.log.info("Running error inject on pe %s function %s",
                               self.pci_device, func)
                 if num_of_miss < 5:
-                    return_code = self.basic_eeh(func)
+                    if self.is_baremetal():
+                        return_code = self.basic_eeh(func, '', '', '', '', '')
+                    else:
+                        return_code = self.basic_eeh(func,
+                                                     self.pci_class_name,
+                                                     self.pci_interface,
+                                                     self.mem_addr,
+                                                     self.mask,
+                                                     self.add_cmd)
                     if return_code == EEH_MISS:
                         num_of_miss += 1
                         self.log.info("number of miss is %d", num_of_miss)
@@ -132,31 +147,48 @@ class PowerNVEEH(Test):
             else:
                 self.fail("PE %s not removed after max hit" % self.pci_device)
 
-    def basic_eeh(self, func):
+    def basic_eeh(self, func, pci_class_name, pci_interface,
+                  pci_mem_addr, pci_mask, add_cmd):
         """
         Injects Error, and checks for PE recovery
         returns True, if recovery is success, else Flase
         """
-        self.clear_dmesg_logs()
-        return_code = self.error_inject(self.addr, self.err, func,
-                                        self.mem_addr, self.mask, self.phb)
+        dmesg.clear_dmesg()
+        if self.is_baremetal():
+            return_code = self.error_inject(func, '', '', self.mem_addr,
+                                            self.mask, '', self.addr,
+                                            self.err, self.phb)
+        else:
+            return_code = self.error_inject(func, pci_class_name,
+                                            pci_interface, pci_mem_addr,
+                                            pci_mask, add_cmd, '', '', '')
         if return_code != EEH_HIT:
             self.log.info("Skipping verification, as command failed")
         if not self.check_eeh_hit():
-            self.log.info("PE %s EEH hit failed", self.pci_device)
+            self.log.info("PE %s EEH hit failed" % self.pci_device)
             return EEH_MISS
-        self.log.info("PE %s EEH hit success", self.pci_device)
-        return EEH_HIT
+        else:
+            self.log.info("PE %s EEH hit success" % self.pci_device)
+            return EEH_HIT
 
-    @classmethod
-    def error_inject(cls, addr, err, func, mem_addr, mask, phb):
+    def error_inject(self, func, pci_class_name, pci_interface, pci_mem_addr,
+                     pci_mask, add_cmd, addr, err, phb):
         """
         Form a command to inject the error
         """
-        cmd = "echo %s:%s:%s:%s:%s > /sys/kernel/debug/powerpc/PCI%s/err_injct \
-               && lspci; echo $?" % (addr, err, func, mem_addr, mask, phb)
-        return int(process.system_output(cmd, ignore_status=True,
-                                         shell=True).decode("utf-8")[-1])
+        if self.is_baremetal():
+            cmd = "echo %s:%s:%s:%s:%s > /sys/kernel/debug/powerpc/PCI%s/err_injct \
+                   && lspci; echo $?" % (addr, err, func, pci_mem_addr, pci_mask, phb)
+            return int(process.system_output(cmd, ignore_status=True,
+                                             shell=True).decode("utf-8")[-1])
+        else:
+            cmd = "errinjct eeh -v -f %s -s %s/%s -a %s -m %s; echo $?"\
+                % (func, pci_class_name, pci_interface, pci_mem_addr, pci_mask)
+            res = process.system_output(cmd, ignore_status=True,
+                                        shell=True).decode("utf-8")
+            if add_cmd:
+                process.run(add_cmd, ignore_status=True, shell=True)
+                return int(res[-1])
 
     def check_eeh_pe_recovery(self):
         """
@@ -191,15 +223,6 @@ class PowerNVEEH(Test):
             return False
 
     @classmethod
-    def clear_dmesg_logs(cls):
-        """
-        Clears dmesg logs, so that functions which uses dmesg
-        gets the latest logs
-        """
-        cmd = "dmesg -C"
-        process.system(cmd, ignore_status=True, shell=True)
-
-    @classmethod
     def check_eeh_hit(cls):
         """
         Function to check if EEH is successfully hit
@@ -207,8 +230,9 @@ class PowerNVEEH(Test):
         tries = 30
         cmd = "dmesg | grep 'EEH: Frozen';echo $?"
         for _ in range(0, tries):
-            if int(process.system_output(cmd, ignore_status=True,
-                                         shell=True).decode("utf-8")[-1]) == 0:
+            res = process.system_output(cmd, ignore_status=True,
+                                        shell=True).decode("utf-8")
+            if int(res[-1]) == 0:
                 return True
             time.sleep(1)
         return False
@@ -220,10 +244,20 @@ class PowerNVEEH(Test):
         """
         tries = 30
         for _ in range(0, tries):
-            cmd = "dmesg | grep 'permanently disabled'; echo $?"
-            if int(process.system_output(cmd, ignore_status=True,
-                                         shell=True).decode("utf-8")[-1]) == 0:
+            cmd = "(dmesg | grep 'permanently disabled'; echo $?)"
+            res = process.system_output(cmd, ignore_status=True,
+                                        shell=True).decode("utf-8")
+            if int(res[-1]) == 0:
                 time.sleep(10)
                 return True
             time.sleep(1)
+        return False
+
+    @staticmethod
+    def is_baremetal():
+        """
+        to check system is bare-metal or not
+        """
+        if 'PowerNV' in genio.read_file("/proc/cpuinfo").strip():
+            return True
         return False
