+++ io/pci/EEH.py.data/EEH.yaml
@@ -1,4 +1,5 @@
 scenario:
-    max_freeze: 1
+    max_freeze: 1 
     function: 4
+    err: 
     pci_device: ""
