--- io/pci/PowerNVEEH.py.data/PowerNVEEH.yaml
@@ -1,5 +1,0 @@
-scenario:
-            max_freeze: 
-            function: 
-            err: 
-            pci_device: ""
